jsondoclint: Extract `Command`/`CommandKind` into its own file
